+++ internal/config/app.go
@@ -7,4 +7,5 @@ type App struct {
 	REST       REST
 	Core       Core
 	Inbox      Inbox
+	Analytics  Analytics
 }