+++ internal/entities/analytics/item.go
@@ -1,0 +1,9 @@
+package analytics
+
+import "time"
+
+type MonthlyActiveUsers struct {
+	PeriodStarted  time.Time
+	ActiveUsers    uint64
+	NewActiveUsers uint64
+}